TST: Split graphics_test to main and others
